Build up inventory items
This commit generally tries to build up the content for dashboard items.

Adds new functions that "hydrate" data models with fields that are
ready-to-use in the Jinja templates:

- insert_github_ref_url (link to the /tree/ view of a
  branch/tag/commit).
- JIRA URL (From the Git ref, matches a ticket branch pattern to
  insert both the ticket name and the ticket's URL in LSST's JIRA.)

Note this is an MVP that works only for single-product repos. This
pattern will need to be revisted for multi-repo products.

The commit also works on styling these dashboard items with flexbox
positioning.