Consolidate recent and state editions to one list
I think it's too confusing to separate recent and state development
editions because what a user really wants to know is what editions have
been merged, and which ones haven't. We don't have that data readily
available though.

Instead, for now these lists can be consolidated. Eventually we can have
a javascript toggle to 'expand' the listing to show older editions that
are not shown by default.