Use box-shadow for links
Following recommendation #43 in the typographychecklist.com, I'm trying
out links based on box-shadows, which produce a more subtle link
indicator that's also easier on descenders. A subtle animation turns the
underline into a background.

Also added a hidden-link class to make links that are normally unstyled,
but have the background on hover. This is useful for titles.

Implemented this link styling on the edition dashboard.