Add JIRA icon to label ticket links
Icon SVG is from https://design.atlassian.com/brand/logo/ but edited to
crop the viewPort to the artwork itself.

Note: needed to specifically tweak the alignment of this icon because
it's not square.

Updated the cheerio gulp usage to remove the st0 class from the JIRA
icon so that it inherits currentFill.