Improve title normalization
The idea is to convert a title "SQR-000: A title" into "A title".
The previous method of doing this with `str.lstrip` wasn't accurate
enough, and this commit uses a regular expression to normalize the
title.

It also splits the document handle extraction (_insert_doc_handle) from
title normalization (_normalize_product_title).

Adds additional testing too.