Make tracked_refs optional
If an edition has a tracked_refs key of None, then the GitHub link is
simply not shown.

An edition can have tracked_refs = None because it uses an alternative
tracking mode.

Also switch to using the edition's `slug` field instead of the
`git_refs` because the slug should always be present.